- Fixing up some compile warnings
git-svn-id: https://dev.chiasoft.net:8443/svn/svn_vana/branches/buis_vana_overhaul@3266 23f3cdc0-e6d1-42be-a0c1-51f4125517bb

+++ src/Common/Item.cpp
@@ -470,7 +470,7 @@ void Item::databaseInsert(soci::session &sql, const vector<ItemDbRecord> &items)
 		location = rec.location;
 		userId = rec.userId;
 		playerId = rec.charId;
-		worldId = rec.worldId;
+		worldId = static_cast<int8_t>(rec.worldId);
 		slot = rec.slot;
 		amount = item->m_amount;
 		itemId = item->m_id;
